Server/requirements.txt
PyYAML
torch
tensorflow
pymongo
colored
pymongo[srv]
flask